Enhance process wrapper by providing user input support
If mutagen process is stuck for more than 5 seconds with no output, it will display it's output to the user so he can enter input, for example when connection to a new ssh host to accept the new key.

It also provide better error messages when mutagen process fails to execute the command.

Close #5
Close #6